got the xcoord working correctly for one character(noface).
only the xcoord works so any height on the page will still trigger a (true) log. But the math checks out, and it is identifying the x correctly. Now just apply the same logic to Y.

I also need to think about how I can clean up this code, so it is more dynamic per game-image.