have a working hhmmss formatted timer/clock going now.
really struggled with setinterval and clearinterval. Has to be done inside the useeffect hook. But where- was the question. I think I have that solved now. Not sure if it's best practice to put the useeffect in the nav, but it works. gameData.hhmmss/time is still working properly and that is in app.js though. So it's time is at least at the top level, and accessible.