lots of changes. Can now end game on a win (instead of only being able to reset game). Created a landing page for game-wins with some basic html on it, and displays the user timestamps.
most important finding, was using useEffect to monitor gameData.characters to find out when all characters have been found. Instead of calling a checkForWin function on setCharacterFound, I just use the useEffect to check gameData.characters being updated. Feels pretty genius lol.

useEffect again the king, able to detect when all characters have been found and sets isGameLive false, and setWin to be true. Changed the children components to render in order of the new variable/state which is 'win/setwin' this way it was easier to create a win game.

debating if I want to change win-screen to be a modal or not. Have it maybe just overlay on top of the other screen (gameArea/ game image in the background of the modal)