preview went from zero to hero real quick once I started copy pasting css. sheesh
copied a lot of the findus css, but hot damn it looks good lol. the preview modal is rounded, the image is not. The characters are displayed, centered and padded.

I think all I need to do is style the start button now, and figure out where i want to put the image info and character from info